requirements.txt
apache-airflow-providers-postgres==6.0.0

